Move json encoding test to test_calculate
